Remove ssh key pair management
I initially set up Terraform to create an SSH keypair for my EC2 instance, and it worked.. but what good is a private key when the value is visible in the TF state file? :facepalm:

Removed the logic to create the key pair and will manage these keys outside of Terraform.