Implement temporary workaround for terraform-docs
terraform-docs currently has a bug that will not allow a subdirectory to generate README files to be places in the root directory. For now we will generate the auto documentation in the tf/ directory and add a link in the root directory until I can find time to tinker with this more.